# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Buildspace API",
    "version": "0.1.0",
    "description": "Separate standalone backend contract. Full request schemas live alongside Hono routes in src/modules."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
...
  "paths": {
...
    "/api/v1/spaces/{slug}/tasks": {
      "get": {
        "summary": "List tasks",
        "tags": [
          "Workspace"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      },
      "post": {
        "summary": "Create tasks",
        "tags": [
          "Workspace"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
...
    "/api/v1/tasks/{taskId}": {
      "get": {
        "summary": "Task detail",
        "tags": [
          "Workspace"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      },
      "patch": {
        "summary": "Update task or board state",
        "tags": [
          "Workspace"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
    "/api/v1/tasks/{taskId}/comments": {
      "post": {
        "summary": "Comment on task",
        "tags": [
          "Workspace"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "503": {
            "description": "Dependency not configured or unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "buildspace_session"
      }
    }
  }
}